* updated all examples according to libfloor v0.4.0 updates:    * made all code work with SDL3 and SDL3_image    * removed all OpenGL code/handling/options    * updated all build.sh files (requires clang 16.0+, use C++23, removed *SSL and OpenAL dependencies, updated warnings, fixed include order)    * updated all CMake files (or added them for projects that didn't have one yet): proper options, initial standalone build support, requires CMake 3.21+, use C++23    * updated all Xcode project (removed various obsolete targets/options/flags, bumped OS requirements to macOS 13.0+ and iOS 16.0+, misc updates, Xcode 15.0+ required now)    * updated config.json    * moved obj_loader.cpp/hpp from common/obj/ to common/obj_loader/    * removed backend specific renderers, only use unified renderers now    * various code updates/modernization/simplifications  * occ:    * added support for disassembling FUBAR files via "--fubar-dis <file.fubar>", implemented in fubar_dis.cpp/hpp    * when disassembling a FUBAR file, this will list all contained targets, their function info and their full disassembly (requires a toolchain)    * pulled out and generalized the function_info dumping from main.cpp and moved it to the new function_info.cpp/hpp    * dropped support for Metal < 3.0 and PTX < 8.0    * general code cleanup  * nbody:    * added initial support for building a standalone binary (requires a yet to be released toolchain to build this)    * added support for building + embedding FUBAR binaries (both nbody itself and libfloor mip-map minify programs)    * together, these enable the building of a fully self-contained binary    * enabled MSAA by default, bumped default body count to 65536    * implemented proper window resize handling

+++ common/camera/camera.hpp
@@ -119,7 +119,7 @@ protected:
 	
 	// hack for mice on os x 10.12+ that don't have proper warp support
 	bool delta_hack { false };
-	int2 last_delta;
+	float2 last_delta;
 	
 	// [right, left, up, down]
 	atomic<bool> key_state[4] { { false }, { false }, { false }, { false } };
